+++ CHANGELOG.md
@@ -21,8 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Source.getPageList`; Suspend alternative to `Source.fetchPageList`.
 - `Source.getFilterList`; Moving up the hierarchy from  `CatalogueSource.getFilterList`.
 - `HttpSource.getHomeUrl`; New method to determine the URL opened when tapping “Open in WebView” in the browse screen.
-- `HttpSource.mangaUpdateRequest`;
-- `HttpSource.mangaUpdateParse`;
 - `HttpSource.getImageUrl`; Suspend alternative to `CatalogueSource.fetchImageUrl`.
 - `HttpSource.getImage`; Suspend alternative to `CatalogueSource.fetchImage`.
 - `SChapter.number` and `SChapter.volume`; Chapter number and volume in string format.
@@ -43,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Source.fetchPageList`; Deprecated in favor of `Source.getPageList`.
 - `HttpSource.fetchImageUrl`; Deprecated in favor of `Source.getImageUrl`.
 - `HttpSource.fetchImage`; Deprecated in favor of `Source.getImage`.
+- `HttpSource.xxxRequest` and `HttpSource.xxxParse`; Deprecated with no replacement.
 - `HttpSource.prepareNewChapter`; Deprecated with no replacement.
 - `rateLimit` and `rateLimitHost`; Deprecated with no replacement.
 - `ParsedHttpSource`; Deprecated with no replacement.